+++ demo/src/components/controls/radio.tsx
@@ -92,8 +92,8 @@ export default function RadioControlView() {
         />
         <wup-radio //
           name="saveUrlRadio"
-          label="With saving to URL (see $options.skey & storage)"
-          skey
+          label="With saving to URL (see $options.storageKey & storage)"
+          storageKey
           storage="url"
           ref={(el) => {
             if (el) {