+++ demo/src/components/controls/select.tsx
@@ -178,8 +178,8 @@ export default function SelectControlView() {
         />
         <wup-select //
           name="saveUrlSelect"
-          label="With saving to URL (see $options.skey & storage)"
-          skey
+          label="With saving to URL (see $options.storageKey & storage)"
+          storageKey
           storage="url"
           ref={(el) => {
             if (el) {