+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Now removing attributes/options always rollbacks to value defined in `$defaults`
   - Method `getAttr` changed to `parseAttr`
 - [SpinElement](src/spinElement.ts). Defaults `fit` & `overflowTarget` = `auto` instead of `null`
+- **Controls**. Option `skey` renamed to `storageKey`
 
 ## 0.8.1 (Aug 29, 2023)